+++ packages/dlt-source-aquabyte/src/dlt_source_aquabyte/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version
 
-from dlt_source_aquabyte.aquabyte import aquabyte_source
+from dlt_source_aquabyte.aquabyte import MAX_WINDOW_DAYS, aquabyte_source
 
 __version__ = version("dlt-source-aquabyte")
-__all__ = ["aquabyte_source"]
+__all__ = ["MAX_WINDOW_DAYS", "aquabyte_source"]